refactor: Move  events to RiskEvaluator to improve performance
Previous code was running the same query 3 times per risk-calculator
enabled.
This modifies it to run the query once, cache in a list, and
re-use for the other  calculators.
